add linux only build tags & remove unused files
