+++ plugins/inputs/slab/slab_nonlinux.go
@@ -1,0 +1,4 @@
+//go:build !linux
+// +build !linux
+
+package slab
