+++ plugins/inputs/slab/slab_sample_config.go
@@ -1,3 +1,6 @@
+//go:build linux
+// +build linux
+
 //go:generate go run ../../../tools/generate_plugindata/main.go
 //go:generate go run ../../../tools/generate_plugindata/main.go --clean
 // DON'T EDIT; This file is used as a template by tools/generate_plugindata
